Folder Structure Fix
Standardized Folder Structure, Defusing Project From Git functionality
