+++ cpp_openGL_drawing_graphics_application/main.cpp
@@ -1,0 +1,14 @@
+#include "Display.h"
+
+#include "Application.h"
+
+int main()
+{
+    Display::init();
+
+    Application app;
+
+    app.runMainGameLoop();
+
+    return 0;
+}
