--- main.cpp
@@ -1,14 +1,0 @@
-#include "Display.h"
-
-#include "Application.h"
-
-int main()
-{
-    Display::init();
-
-    Application app;
-
-    app.runMainGameLoop();
-
-    return 0;
-}
